added mnist gcnn

+++ lib/graph/preprocess.py
@@ -3,8 +3,7 @@ import scipy.sparse as sp
 
 
 def preprocess_adj(adj):
-    adj = adj + sp.eye(adj.shape[0])
-
+    adj = adj + sp.eye(adj.shape[0], dtype=adj.dtype)
     deg = np.array(adj.sum(1)).flatten()
 
     with np.errstate(divide='ignore'):
@@ -12,4 +11,5 @@ def preprocess_adj(adj):
         deg[np.isinf(deg)] = 0
 
     deg = sp.diags(deg)
+
     return deg.dot(adj).dot(deg)
